+++ db/migrations/0018-20170325-admin_user_tfa.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `admin_users` ADD `tfa_secret` VARCHAR(255) NOT NULL AFTER `user_password`;